requirements.txt
python==3.8
black==19.10b0
fastapi==0.61.1
jsonpickle==1.4.1
pylint==2.6.0
uvicorn==0.12.1
pythonocc-core==7.4.0
pytest==6.2.1